DebugInfo: Get correct file path for substituted methods
When computing the full file path for the source file of a method check
to see if this is a substitution method and get the source file
containing it instead of the source file containing the original method.

So if we have Foo.java that defines Foo.m which substitutes Bar.n which
is defined in Bar.java this patch will associate code segments of Bar.n
with Foo.java instead of Bar.java.